Frame skipping moved to JAABA thread. Work in progress

+++ src/gui/image_grabber.hpp
@@ -84,7 +84,7 @@ namespace bias
             unsigned int cameraNumber_;
             bool testConfigEnabled_;
             string trial_num;
-            float fstfrmtStampRef_;
+            uint64_t fstfrmtStampRef_;
 
             unsigned int partnerCameraNumber_;
             uInt32 read_buffer_ = 0, read_ondemand_ = 0;
